+++ module/Shared/src/Infrastructure/Clock/FrozenClock.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Shared\Infrastructure\Clock;
+
+use DateTimeImmutable;
+use ExtendsSoftware\ExaPHPExample\Shared\Application\Clock\ClockInterface;
+
+final readonly class FrozenClock implements ClockInterface
+{
+    public function __construct(private DateTimeImmutable $time) {}
+
+    public function now(): DateTimeImmutable
+    {
+        return $this->time;
+    }
+}